[SAFE-411] virtual-assistant directory moved + text messages sending managed
