+++ projects/back-office/src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { MsalGuard } from '@azure/msal-angular';
 import { AccessGuard } from './guards/access.guard';
-import {VirtualAssistantComponent} from './dashboard/pages/virtual-assistant/virtual-assistant.component';
+import {VirtualAssistantComponent} from './virtual-assistant/virtual-assistant.component';
 
 const routes: Routes = [
   {
@@ -35,12 +35,10 @@ const routes: Routes = [
       },
       {
           path: 'va',
-          // component: VaComponent
           children: [
           {
             path: ':id',
-            // component: VirtualAssistantComponent
-            loadChildren: () => import('./dashboard/pages/virtual-assistant/virtual-assistant.module')
+            loadChildren: () => import('./virtual-assistant/virtual-assistant.module')
               .then(m => m.VirtualAssistantModule),
           }
         ]
