+++ projects/back-office/src/app/virtual-assistant/models/choices.model.ts
