+++ projects/back-office/src/app/virtual-assistant/models/message.model.ts
