+++ projects/back-office/src/app/virtual-assistant/models/user.model.ts
