+++ projects/back-office/src/app/virtual-assistant/va-conversation/conversation-footer/conversation-footer.component.html
