+++ projects/back-office/src/app/virtual-assistant/va-conversation/conversation-header/conversation-header.component.spec.ts
