+++ projects/back-office/src/app/virtual-assistant/va-conversation/conversation-message/conversation-message.component.ts
