+++ projects/back-office/src/app/virtual-assistant/va-conversation/va-conversation.component.spec.ts
