+++ projects/back-office/src/app/virtual-assistant/virtual-assistant.component.scss
